--- a/dprint.json
+++ b/dprint.json
@@ -1,0 +1,30 @@
+{
+  "prettier": {
+    "arrowParens": "avoid",
+    "trailingComma": "all",
+    "singleQuote": true
+  },
+  "includes": [
+    "**/*.{ts,tsx,js,jsx,cjs,mjs}"
+  ],
+  "excludes": [
+    "**/node_modules",
+    "**/*-lock.json",
+    ".yarn",
+    "bundles",
+    "compiled",
+    "__generated",
+    "dist",
+    "moddable",
+    "public",
+    "vendor",
+    // subrepos
+    "/packages/xsnap/moddable",
+    "/packages/xsnap/xsnap-native"
+  ],
+  "plugins": [
+    "https://plugins.dprint.dev/toml-0.5.4.wasm",
+    "https://plugins.dprint.dev/dockerfile-0.3.0.wasm",
+    "https://plugins.dprint.dev/prettier-0.58.2.json@2fd8cc9c9773364848dc93735b01a00d55fcfa78a56a9230b4b63c014964c82d"
+  ]
+}