--- a/packages/orchestration/tsup.config.ts
+++ b/packages/orchestration/tsup.config.ts
@@ -13,14 +13,5 @@ export default defineConfig({
   dts: true,
   bundle: true,
   platform: 'neutral',
-  external: [
-    'stream',
-    'crypto',
-    'http',
-    'https',
-    'zlib',
-    'net',
-    'tls',
-    'url',
-  ],
+  external: ['stream', 'crypto', 'http', 'https', 'zlib', 'net', 'tls', 'url'],
 });