## Description

Runs prettier check and format within [dprint](https://dprint.dev/) for parallelization and caching (see https://david.deno.dev/posts/faster-prettier-with-dprint/ )

**Before**
```
❯ yarn format
yarn run v1.22.19
$ yarn prettier --write .github golang packages
$ /opt/agoric/agoric-sdk/node_modules/.bin/prettier --write .github golang packages
…
✨  Done in 8.57s.
```

```
❯ yarn lint:format
yarn run v1.22.19
$ yarn prettier --check .github golang packages
$ /opt/agoric/agoric-sdk/node_modules/.bin/prettier --check .github golang packages
…
✨  Done in 8.35s.
```

**After**
```
❯ yarn format
yarn run v1.22.19
$ dprint fmt
✨  Done in 0.30s.
```

```
❯ yarn lint:format
yarn run v1.22.19
$ dprint check
✨  Done in 0.34s.
```

### Security Considerations

--

### Scaling Considerations

--

### Documentation Considerations

--

### Testing Considerations